refactor(s3s-sigv4): extract crypto helpers into a crypto module
Move the SHA-256 helpers (hex_bytes32, sha256/sha256_chunk with their
openssl variants, hex_sha256, hex_sha256_chunk, hmac_sha256) out of
methods.rs into the existing pub(crate) crypto module, alongside
is_sha256_checksum. methods.rs now only carries canonicalization and
signing; the openssl feature cfg is concentrated in one file. A
self-test pins the canonical hashes used by the AWS examples.
